+++ functions/src/functions/line.ts
@@ -1,51 +1,7 @@
 import * as functions from 'firebase-functions'
 import * as utils from '../stripe/utils'
-import * as https from 'https'
-import * as url from 'url';
-import * as admin from 'firebase-admin';
-
-const request = (_url: string, _options: any, postData?: any): Promise<any> => {
-  const parsedURL = url.parse(_url);
-  const options = Object.assign({
-    hostname: parsedURL.host,
-    port: 443,
-    path: parsedURL.pathname,
-    method: "GET"
-  }, _options)
-  return new Promise((resolve, reject) => {
-    const req = https.request(options, (res) => {
-      res.setEncoding('utf8');
-      let body = "";
-      res.on('data', (chunk) => {
-        body += chunk;
-      });
-      res.on('end', () => {
-        resolve(JSON.parse(body))
-      })
-    })
-    req.on('error', (e) => {
-      console.error(e);
-      reject({ error: e });
-    });
-    if (postData) {
-      req.write(postData)
-    }
-    req.end();
-  })
-}
-
-const postForm = (_url: string, params: any): Promise<any> => {
-  const postData = Object.keys(params).map(key => {
-    return key + "=" + encodeURIComponent(params[key]);
-  }).join("&");
-  return request(_url, {
-    method: "POST",
-    headers: {
-      'Content-Type': 'application/x-www-form-urlencoded',
-      'Content-Length': Buffer.byteLength(postData)
-    }
-  }, postData)
-}
+import * as netutils from '../lib/netutils'
+import * as admin from 'firebase-admin'
 
 export const validate = async (db: FirebaseFirestore.Firestore, data: any, context: functions.https.CallableContext) => {
   const { code, redirect_uri, client_id } = data;
@@ -62,12 +18,22 @@ export const validate = async (db: FirebaseFirestore.Firestore, data: any, conte
 
   try {
     // access_token, id_token, expires_in, refresh_token, scope, token_type
-    const result = await postForm("https://api.line.me/oauth2/v2.1/token", params)
+    const result = await netutils.postForm("https://api.line.me/oauth2/v2.1/token", params)
+    if (!result.id_token) {
+      throw new functions.https.HttpsError('invalid-argument',
+        'Validation failed.', { params: result }
+      )
+    }
     // amr, aud, exp, iat, iss, name, sub
-    const verified = await postForm('https://api.line.me/oauth2/v2.1/verify', {
+    const verified = await netutils.postForm('https://api.line.me/oauth2/v2.1/verify', {
       id_token: result.id_token,
       client_id
     })
+    if (!verified.sub) {
+      throw new functions.https.HttpsError('invalid-argument',
+        'Verification failed.', { params: verified }
+      )
+    }
     const customeToken = await admin.auth().createCustomToken(`line:${verified.sub}`)
 
     return { result, verified, customeToken };
